+++ Logic/Models/Item.cs
@@ -13,14 +13,14 @@ namespace Logic.Models
         public string ProductName { get; set; }
         public string Description { get; set; }
         public int VAT { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Amount { get; set; }
         public bool InStock { get; set; }
 
         public Item()
         { }
 
-        public Item(string id, string productName, string description, int vat, double price, int amount, bool instock)
+        public Item(string id, string productName, string description, int vat, decimal price, int amount, bool instock)
         {
             ProductName = productName;
             ItemID = id;
